saisFunction/commands.py
# Import any necessary libraries or modules here


# Define functions to handle different commands
from saisFunction.database import insert_interaction


def greet_user():
    return "Hello! How can I assist you today?"

def get_weather():
    # Implement code to retrieve weather information from a source (e.g., an API)
    # You can use external libraries like requests to make API requests
    # Return the weather information as a response
    return "The weather today is sunny with a high of 28°C."

def search_wikipedia():
    # Implement code to search Wikipedia for a user's query
    # You can use libraries like wikipedia-api to interact with Wikipedia
    # Return a summary or relevant information as a response
    return "Wikipedia search result: [Your search result here]"

# Define a dictionary of predefined commands and their corresponding functions
# You can expand this dictionary with more commands and functions as needed
COMMANDS = {
    "greet": greet_user,
    "get_weather": get_weather,
    "search_wikipedia": search_wikipedia,
    # Add more commands here
}

# Define a function to process user input and return a response
def process_user_input(user_input, conn):
    # Split the user input into words or tokens
    tokens = user_input.split()

    # Check if the first token (word) is a valid command
    if tokens[0] in COMMANDS:
        # Call the corresponding command function
        response = COMMANDS[tokens[0]]()
    else:
        # Handle unrecognized commands or provide a default response
        response = "I'm sorry, I don't understand that command."

    # Insert the interaction into the database
    insert_interaction(conn, user_input, response)
    
    return response

# Main function to test the commands
if __name__ == "__main__":
    while True:
        user_input = input("Enter a command: ")
        response = process_user_input(user_input)
        print(response)
